packages_py/static_config_property_management/pyproject.toml
[tool.poetry]
name = "static-config-property-management"
version = "0.1.0"
description = "A module for loading and managing static YAML configuration properties"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "static_config", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
pydantic = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
pytest-cov = "^4.1"
mypy = "^1.8"
ruff = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = []

[tool.coverage.run]
source = ["src/static_config"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
